stopjudol/client/requirements.txt
PyQt6>=6.4.0
google-api-python-client>=2.70.0
google-auth-oauthlib>=1.0.0
google-auth>=2.16.0
keyring>=23.13.1
requests>=2.28.2
jsonrpcclient>=4.0.0
